config: use a single directory variable to represent $lib/X11/rstart
There are numerous places where the rstart lib directory is needed.
One of the 4 strings is used, all have the same value:

rstart_serverdir
$(libdir)/X11/rstart
configdir
DATA_DIR

There are no configure options to let user assign a different directory
for one of the places where the value is needed.

The patch assigns a new value called rstartdir to represent the root
dir from where files and other directories are created. Because it
ends with "dir" it can be used as a prefix to primary variables
such as PROGRAMS or DATA.

Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>