config: add links from x and x11 from @Aliases
This was part of the FIXME code removed earlier.
The information was read from the @Aliases file, but it
has not changed for the past 8 years. If it does change,
the makefile will need to be updated.

The links are broken in 1.0.2 and restored in 1.0.3

Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Using Automake, these links are now properly removed when uninstalling.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>